refactor: move db module into db/ package
